requirements.txt
# TradeFlow - Secure Document Upload Portal
# Python Dependencies

# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Google Drive API
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Utilities
python-dotenv==1.0.0